CR for Hugo
