+++ apps/festival/festival-e2e/src/support/pages/dashboard/index.ts
@@ -1,0 +1,9 @@
+import EventEditPage from './EventEditPage';
+import EventPage from './EventPage';
+import FestivalDashboardHomePage from './FestivalDashboardHomePage';
+
+export {
+  EventEditPage,
+  EventPage,
+  FestivalDashboardHomePage
+};
