+++ apps/festival/festival-e2e/src/support/pages/marketplace/index.ts
@@ -1,0 +1,5 @@
+import FestivalMarketpalceHomepage from './FestivalMarketplaceHomePage';
+
+export {
+  FestivalMarketpalceHomepage
+}
