+++ libs/e2e/src/lib/utils/functions.ts
@@ -39,3 +39,11 @@ export function assertUploadStatus(content: string, testId: string) {
 let currentID = 0;
 export const randomID = (): string => (`${new Date().toISOString()}-${currentID++}`);
 export const createFakeScript = (title: string): any => cy.task('random:pdf', title);
+
+export function formatAmPm(date: Date, toLowerCase?: boolean) {
+  let hours = date.getHours();
+  const amPM = hours >= 12 ? 'PM' : 'AM';
+  hours = hours % 12;
+  hours = hours ? hours : 12; // the hour 0 should be 12
+  return toLowerCase ? hours + ' ' + amPM.toLowerCase() : hours + ' ' + amPM;
+}
